Fix build errors

+++ v2/tests/log_context_property_type_ascii_char_ptr_int/log_context_property_type_ascii_char_ptr_int.c
@@ -55,7 +55,7 @@ static void ascii_char_ptr_to_string_with_NULL_buffer_and_zero_buffer_length_ret
     int result = LOG_CONTEXT_PROPERTY_TYPE_IF_IMPL(ascii_char_ptr).to_string("cucu", NULL, 0);
 
     // assert
-    POOR_MANS_ASSERT(result == strlen("cucu"));
+    POOR_MANS_ASSERT(result == (int)strlen("cucu"));
 }
 
 static void ascii_char_ptr_to_string_copies_the_string(void)
